tests/knowledge/rules/aws/non_context_aware/log_validation_rules/test_ensure_api_gw_xray_tracing_enabled_rule.py
import unittest

from cloudrail.dev_tools.rule_test_utils import create_empty_entity
from cloudrail.knowledge.context.aws.apigateway.api_gateway_stage import ApiGatewayStage
from cloudrail.knowledge.context.aws.apigateway.rest_api_gw import RestApiGw
from cloudrail.knowledge.context.environment_context import EnvironmentContext
from cloudrail.knowledge.rules.aws.non_context_aware.log_validation_rules.ensure_api_gw_xray_tracing_enabled_rule import \
    EnsureApiGwXrayTracingEnabledRule
from cloudrail.knowledge.rules.base_rule import RuleResultType


class TestEnsureApiGwXrayTracingEnabledRule(unittest.TestCase):
    def setUp(self):
        self.rule = EnsureApiGwXrayTracingEnabledRule()

    def test_non_car_api_gateway_xray_tracing_enabled_fail(self):
        # Arrange
        rest_api_gw: RestApiGw = create_empty_entity(RestApiGw)
        rest_api_stage: ApiGatewayStage = create_empty_entity(ApiGatewayStage)
        rest_api_stage.xray_tracing_enabled = False
        rest_api_gw.api_gw_stages = [rest_api_stage]
        context = EnvironmentContext(rest_api_gw=[rest_api_gw], rest_api_stages=[rest_api_stage])
        # Act
        result = self.rule.run(context, {})
        # Assert
        self.assertEqual(RuleResultType.FAILED, result.status)
        self.assertEqual(1, len(result.issues))

    def test_non_car_api_gateway_xray_tracing_enabled__tracing_enabled__pass(self):
        # Arrange
        rest_api_gw: RestApiGw = create_empty_entity(RestApiGw)
        rest_api_stage: ApiGatewayStage = create_empty_entity(ApiGatewayStage)
        rest_api_stage.xray_tracing_enabled = True
        rest_api_gw.api_gw_stages = [rest_api_stage]
        context = EnvironmentContext(rest_api_gw=[rest_api_gw], rest_api_stages=[rest_api_stage])
        # Act
        result = self.rule.run(context, {})
        # Assert
        self.assertEqual(RuleResultType.SUCCESS, result.status)
        self.assertEqual(0, len(result.issues))

    def test_non_car_api_gateway_xray_tracing_enabled__no_stage__pass(self):
        # Arrange
        rest_api_gw: RestApiGw = create_empty_entity(RestApiGw)
        rest_api_stage: ApiGatewayStage = create_empty_entity(ApiGatewayStage)
        rest_api_stage.xray_tracing_enabled = True
        rest_api_gw.api_gw_stages = []
        context = EnvironmentContext(rest_api_gw=[rest_api_gw], rest_api_stages=[rest_api_stage])
        # Act
        result = self.rule.run(context, {})
        # Assert
        self.assertEqual(RuleResultType.SUCCESS, result.status)
        self.assertEqual(0, len(result.issues))
